[Net] Move CBanDB and CAddrDB out of net.h/cpp
